docs: update filename option default in examples
Update the `filename` option in the examples across the documentation
to clarify that it is optional and defaults to `camelCase`. This
ensures users are aware that if unspecified, import declarations such
as `DatePicker` will automatically convert to `datePicker` when using
`camelCase`. This change improves documentation accuracy and helps
users better understand default behaviors.